+++ DTCD-HeaderPanel/src/Plugin.js
@@ -9,12 +9,22 @@ import {
   StyleSystemAdapter,
   RouteSystemAdapter,
   AppGUISystemAdapter,
+  NotificationSystemAdapter,
 } from './../../DTCD-SDK/index';
 
 export class Plugin extends AppPanelPlugin {
   #vue;
   #workspaceSystem;
-  #settings;
+
+  #settings = {
+    showPageTitle: false,
+    showAddPanelButton: false,
+    showBackButton: false,
+    showSettingsButton: false,
+    showWorkspaceSettings: false,
+    showPanelSelect: false,
+    notificationPosition: 'right top',
+  };
 
   static getRegistrationMeta() {
     return pluginMeta;
@@ -28,17 +38,20 @@ export class Plugin extends AppPanelPlugin {
     const styleSystem = new StyleSystemAdapter('0.4.0');
     const router = new RouteSystemAdapter('0.1.0');
     const appGUI = new AppGUISystemAdapter('0.1.0');
+    const notificationSystem = new NotificationSystemAdapter('0.1.0');
+
+    eventSystem.registerPluginInstance(this, [
+      'newNotify',
+      'removeNotify',
+      'clearNotifyList',
+    ]);
+
+    const guidNotifySystem = this.getGUID(this.getSystem('NotificationSystem', '0.1.0'));
+    eventSystem.subscribe(guidNotifySystem, 'newNotify', guid, 'onNewNotify');
+    eventSystem.subscribe(guidNotifySystem, 'removeNotify', guid, 'onRemoveNotify');
+    eventSystem.subscribe(guidNotifySystem, 'clearNotifyList', guid, 'onClearNotifyList');
 
     const VueJS = this.getDependence('Vue');
-
-    this.#settings = {
-      showPageTitle: false,
-      showAddPanelButton: false,
-      showBackButton: false,
-      showSettingsButton: false,
-      showWorkspaceSettings: false,
-      showPanelSelect: false,
-    };
 
     const data = {
       plugin: this,
@@ -49,38 +62,71 @@ export class Plugin extends AppPanelPlugin {
       styleSystem,
       router,
       appGUI,
+      notificationSystem,
+      settings: this.#settings,
     };
 
     this.#vue = new VueJS.default({
       data: () => data,
       render: h => h(PluginComponent),
+      methods: {
+        setPanelSettings(settings) {
+          for (let key in settings) {
+            this.settings[key] = settings[key];
+          }
+        },
+      }
     }).$mount(selector);
   }
 
+  onNewNotify(obj) {
+    this.#vue.$emit('onNewNotify', obj)
+  }
+
+  onRemoveNotify(id) {
+    this.#vue.$emit('onRemoveNotify', id)
+  }
+
+  onClearNotifyList(id) {
+    this.#vue.$emit('onClearNotifyList')
+  }
+
   setPluginConfig(configuration) {
-    this.resetSettings();
     for (let setting in configuration) {
       if (this.#settings.hasOwnProperty(setting)) {
         this.#settings[setting] = configuration[setting];
       }
     }
+    this.#vue.setPanelSettings(this.#settings);
     this.#vue.$children[0].$children[1].setPanelSettings(this.#settings);
   }
 
-  resetSettings() {
-    this.#settings = {
-      showPageTitle: false,
-      showAddPanelButton: false,
-      showBackButton: false,
-      showSettingsButton: false,
-      showWorkspaceSettings: false,
-      showPanelSelect: false,
-    };
+  getPluginConfig() {
+    return { ...this.#settings };
   }
 
-  getPluginConfig() {}
+  setFormSettings(config) {
+    this.setPluginConfig(config);
+  }
 
-  setFormSettings() {}
-
-  getFormSettings() {}
+  getFormSettings() {
+    return {
+      fields: [
+        {
+          component: 'title',
+          innerText: 'Уведомления',
+        },
+        {
+          component: 'select',
+          propName: 'notificationPosition',
+          options: [
+            { label: 'Top right', value: 'top right' },
+            { label: 'Bottom right', value: 'bottom right' },
+            { label: 'Top left', value: 'top left' },
+            { label: 'Bottom left', value: 'bottom left' },
+          ],
+        },
+      ]
+    }
+  }
 }