+++ CHANGELOG.md
@@ -5,19 +5,32 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### [0.4.0]
+## [0.5.0]
+
+### Added
+
+- request for user avatar compression
+- notification bell with notification list
+- text if there are no notifications
+- if you hover the mouse over a floating notification, it will not hide automatically
+
+### Changed
+
+- SDK version to 0.2.0-master-0002
+
+## [0.4.0]
 
 ### Changed
 
 - API prefix
 
-### [0.3.0]
+## [0.3.0]
 
 ### Fixed
 
 - version select
 
-### [0.2.0]
+## [0.2.0]
 
 ### Added